added utility functions to test_utils

+++ include/test_utils.h
@@ -5,11 +5,12 @@
 #ifndef SORTING_SRC_TEST_UTILS_H_
 #define SORTING_SRC_TEST_UTILS_H_
 
+#include <iostream>
 #include <vector>
 
 using namespace std;
 
 void display_array(vector<int> arr);
-vector<int> gen_random_arr(int size);
+vector<int> gen_random_arr(long size);
 
 #endif //SORTING_SRC_TEST_UTILS_H_
